Fix integration tests
Originally, the integration tests file, list, and list_and_file would
check if a file was backed up correctly by opening the directory,
putting the filenames into a vector, and asserting from the indices of
the vector. This was unreliable, so instead of assigning a file to a
variable, and asserting that the filename is equal to the filename from
the index, it now asserts that the filename is contained within the vector.
The tar integration test was fixed also. The problem was that the
default directory was hard coded which means the build would only pass
once a year :).